requirements.txt
pyautogen==0.1.5
docker==6.1.3
openai==0.28.0

# to use docker install wsl in windows then docker desktop and run docker desktop
# "wsl --install in powershell to install wsl